extraindo em módulos pt1
