+++ space-invaders/src/game/resources.rs
@@ -2,22 +2,6 @@ use bevy::prelude::Resource;
 
 use super::systems::PLAYER_HP;
 
-#[derive(Resource, Default, Clone, Copy, Debug)]
-pub enum AlienDirection {
-    #[default]
-    Right,
-    Left,
-}
-
-impl AlienDirection {
-    pub fn toggle(&mut self) {
-        match self {
-            AlienDirection::Right => *self = AlienDirection::Left,
-            AlienDirection::Left => *self = AlienDirection::Right,
-        };
-    }
-}
-
 #[derive(Resource, Default)]
 pub struct Score(pub usize);
 
@@ -56,5 +40,8 @@ impl Lives {
         if self.0 > 0 {
             self.0 -= 1
         }
+    }
+    pub fn zero(&mut self) {
+        self.0 = 0
     }
 }
